reporter/pyproject.toml
[project]
name = "reporter"
version = "0.1.0"
description = "PostgreSQL MCP Server - 数据查询与统计推送服务"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.27.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
reporter-mcp = "reporter.server:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling]
package-dir = "src"
packages = ["reporter"]